chatbot/requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
qdrant-client==1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-multipart==0.0.6
httpx==0.26.0
tiktoken==0.5.2
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.50.0
qdrant-client==1.11.0
python-dotenv==1.0.1
EOF
